Utils.h and utils.c moved to utils folder
