+++ client/netdb.c
@@ -13,7 +13,7 @@
 #include "init.h"
 #include "block.h"
 #include "sync.h"
-#include "utils.h"
+#include "../utils/utils.h"
 
 #define MAX_SELECTED_HOSTS  64
 #define MAX_BLOCKED_IPS     64
