+++ client/terminal.c
@@ -33,7 +33,7 @@ const uint32_t LOCAL_HOST_IP = 0x7f000001; // 127.0.0.1
 const uint32_t APPLICATION_DOMAIN_PORT = 7676;
 #endif
 
-#include "utils.h"
+#include "../utils/utils.h"
 
 int terminal(void)
 {
