+++ win/xdaglib/xdaglib.vcxproj.filters
@@ -34,6 +34,9 @@
     <Filter Include="win\sys">
       <UniqueIdentifier>{874b2b63-f358-4dea-b109-a761875c2512}</UniqueIdentifier>
     </Filter>
+    <Filter Include="utils">
+      <UniqueIdentifier>{71f09615-b936-42fb-b99c-61c593b206be}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\..\client\address.c">
@@ -134,6 +137,9 @@
     </ClCompile>
     <ClCompile Include="..\..\client\init.c">
       <Filter>client</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\utils\utils.c">
+      <Filter>utils</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -289,6 +295,9 @@
     </ClInclude>
     <ClInclude Include="..\..\client\init.h">
       <Filter>client</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\utils\utils.h">
+      <Filter>utils</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
